Run all the tests successfully
