+++ jslack-api-client/src/test/java/test_with_remote_apis/methods_admin_api/AdminApiTest.java
@@ -81,6 +81,7 @@ public class AdminApiTest {
         }
     }
 
+    @Ignore // Doesn't work as of Jan 3, 2020
     @Test
     public void changeSharedChannels() throws Exception {
         if (userToken != null && orgAdminToken != null && sharedChannelId != null) {
