+++ jslack-api-client/src/test/java/util/sample_json_generation/JsonDataRecorder.java
@@ -2,6 +2,9 @@ package util.sample_json_generation;
 
 import com.github.seratch.jslack.SlackConfig;
 import com.github.seratch.jslack.api.model.Conversation;
+import com.github.seratch.jslack.api.model.FileComment;
+import com.github.seratch.jslack.api.model.Message;
+import com.github.seratch.jslack.api.model.admin.AppRequest;
 import com.github.seratch.jslack.api.scim.model.User;
 import com.github.seratch.jslack.api.status.v2.model.SlackIssue;
 import com.github.seratch.jslack.common.json.GsonFactory;
@@ -48,8 +51,7 @@ public class JsonDataRecorder {
             // given body is not in JSON format
             return;
         }
-        JsonParser jsonParser = new JsonParser();
-        JsonElement newJson = jsonParser.parse(body);
+        JsonElement newJson = JsonParser.parseString(body);
 
         String existingJson = null;
         if (path.startsWith("/scim")) {
@@ -71,7 +73,7 @@ public class JsonDataRecorder {
                 existingJson = "{}";
             }
         }
-        JsonElement jsonElem = jsonParser.parse(existingJson);
+        JsonElement jsonElem = JsonParser.parseString(existingJson);
         JsonObject jsonObj = jsonElem.isJsonObject() ? jsonElem.getAsJsonObject() : null;
 
         if (newJson.isJsonObject()) {
@@ -89,7 +91,7 @@ public class JsonDataRecorder {
 
         if (jsonElem.isJsonObject() && jsonObj != null) {
             for (Map.Entry<String, JsonElement> entry : jsonObj.entrySet()) {
-                scanToNormalizeStringValues(path, jsonObj, entry.getKey(), entry.getValue());
+                scanToNormalizeValues(path, jsonObj, entry.getKey(), entry.getValue());
             }
 
             if (path.startsWith("/scim")) {
@@ -122,7 +124,7 @@ public class JsonDataRecorder {
             }
         } else if (jsonElem.isJsonArray()) {
             JsonArray jsonArray = jsonElem.getAsJsonArray();
-            scanToNormalizeStringValues(path,null, null, jsonArray);
+            scanToNormalizeValues(path,null, null, jsonArray);
             existingJson = gson().toJson(jsonArray);
             Path filePath = new File(toMaskedFilePath(path)).toPath();
             Files.createDirectories(filePath.getParent());
@@ -208,7 +210,8 @@ public class JsonDataRecorder {
 
     private MergeJsonBuilder.ConflictStrategy CONFLICT_STRATEGY = MergeJsonBuilder.ConflictStrategy.PREFER_FIRST_OBJ;
 
-    private void scanToNormalizeStringValues(String path, JsonElement parent, String name, JsonElement element) {
+    private void scanToNormalizeValues(String path, JsonElement parent, String name, JsonElement element) {
+        Gson gson = GsonFactory.createSnakeCase();
         if (element.isJsonArray()) {
             JsonArray array = element.getAsJsonArray();
             if (name != null && name.equals("attachments")) {
@@ -225,6 +228,16 @@ public class JsonDataRecorder {
                 for (JsonElement block : Json.Blocks) {
                     array.add(block);
                 }
+            } else if (name != null && name.equals("replies")) {
+                for (int idx = 0; idx < array.size(); idx++) {
+                    array.remove(idx);
+                }
+                array.add(gson.toJsonTree(ObjectInitializer.initProperties(new Message.MessageRootReply())));
+            } else if (name != null && name.equals("comments")) {
+                for (int idx = 0; idx < array.size(); idx++) {
+                    array.remove(idx);
+                }
+                array.add(gson.toJsonTree(ObjectInitializer.initProperties(new FileComment())));
             }
             if (array.size() == 0) {
                 List<String> addressNames = Arrays.asList("from", "to", "cc");
@@ -233,16 +246,22 @@ public class JsonDataRecorder {
                     address.setAddress("");
                     address.setName("");
                     address.setOriginal("");
-                    JsonElement elem = GsonFactory.createSnakeCase().toJsonTree(address);
+                    JsonElement elem = gson.toJsonTree(address);
                     array.add(elem);
                 } else if (path.equals("/api/users.conversations") && name.equals("channels")) {
-                    array.add(GsonFactory.createSnakeCase().toJsonTree(ObjectInitializer.initProperties(new Conversation())));
+                    array.add(gson.toJsonTree(ObjectInitializer.initProperties(new Conversation())));
+                } else if (name.equals("app_requests")) {
+                    array.add(gson.toJsonTree(ObjectInitializer.initProperties(new AppRequest())));
+                } else if (name.equals("replies")) {
+                    array.add(gson.toJsonTree(ObjectInitializer.initProperties(new Message.MessageRootReply())));
+                } else if (name.equals("comments")) {
+                    array.add(gson.toJsonTree(ObjectInitializer.initProperties(new FileComment())));
                 } else if (name.equals("active_incidents")) {
                     SlackIssue slackIssue = new SlackIssue();
                     slackIssue.setNotes(Arrays.asList(ObjectInitializer.initProperties(new SlackIssue.Note())));
                     slackIssue.setServices(Arrays.asList(""));
                     slackIssue = ObjectInitializer.initProperties(slackIssue);
-                    array.add(GsonFactory.createSnakeCase().toJsonTree(slackIssue));
+                    array.add(gson.toJsonTree(slackIssue));
                 } else {
                     array.add(""); // in most cases, empty array can have string values
                 }
@@ -258,14 +277,14 @@ public class JsonDataRecorder {
                     }
                 } else {
                     for (JsonElement child : array) {
-                        scanToNormalizeStringValues(path, array, null, child);
+                        scanToNormalizeValues(path, array, null, child);
                     }
                     if (array.size() >= 2) {
                         for (int idx = 1; idx < array.size(); idx++) {
                             JsonElement elem = array.get(idx);
                             if (elem.isJsonArray()) {
                                 for (JsonElement child : elem.getAsJsonArray()) {
-                                    scanToNormalizeStringValues(path, elem, null, child);
+                                    scanToNormalizeValues(path, elem, null, child);
                                 }
                             } else {
                                 try {
@@ -307,7 +326,7 @@ public class JsonDataRecorder {
                 }
             }
             for (Map.Entry<String, JsonElement> entry : element.getAsJsonObject().entrySet()) {
-                scanToNormalizeStringValues(path, element, entry.getKey(), entry.getValue());
+                scanToNormalizeValues(path, element, entry.getKey(), entry.getValue());
             }
         } else if (element.isJsonNull()) {
             return;
