+++ json-logs/samples/api/admin.apps.requests.list.json
@@ -1,7 +1,48 @@
 {
   "ok": false,
   "app_requests": [
-    ""
+    {
+      "id": "",
+      "app": {
+        "id": "",
+        "name": "",
+        "description": "",
+        "help_url": "",
+        "privacy_policy_url": "",
+        "app_homepage_url": "",
+        "app_directory_url": "",
+        "is_app_directory_approved": false,
+        "is_internal": false,
+        "additional_info": "",
+        "icons": {
+          "image_32": "",
+          "image_36": "",
+          "image_48": "",
+          "image_64": "",
+          "image_72": "",
+          "image_96": "",
+          "image_128": "",
+          "image_192": "",
+          "image_512": "",
+          "image_1024": ""
+        }
+      },
+      "user": {
+        "id": "",
+        "name": "",
+        "email": ""
+      },
+      "team": {
+        "id": "",
+        "name": "",
+        "domain": ""
+      },
+      "previous_resolution": {
+        "status": ""
+      },
+      "message": "",
+      "date_created": 123
+    }
   ],
   "response_metadata": {
     "next_cursor": ""
