+++ json-logs/samples/api/chat.postMessage.json
@@ -46,6 +46,11 @@
           },
           {
             "type": "",
+            "text": "",
+            "emoji": false
+          },
+          {
+            "type": "",
             "fallback": "",
             "text": {
               "type": "",
@@ -162,8 +167,11 @@
     },
     "attachments": [
       {
-        "msg_subtype": "",
+        "text": "",
+        "footer": "",
+        "id": 12345,
         "fallback": "",
+        "msg_subtype": "",
         "callback_id": "",
         "color": "",
         "pretext": "",
@@ -178,7 +186,6 @@
         "author_subname": "",
         "channel_id": "",
         "channel_name": "",
-        "id": 12345,
         "bot_id": "",
         "indent": false,
         "is_msg_unfurl": false,
@@ -188,7 +195,6 @@
         "app_unfurl_url": "",
         "title": "",
         "title_link": "",
-        "text": "",
         "fields": [
           {
             "title": "",
@@ -206,7 +212,6 @@
         "video_html": "",
         "video_html_width": 12345,
         "video_html_height": 12345,
-        "footer": "",
         "footer_icon": "",
         "ts": "",
         "mrkdwn_in": [
