+++ json-logs/samples/api/chat.update.json
@@ -31,6 +31,16 @@
         "elements": [
           {
             "type": "",
+            "action_id": "",
+            "text": {
+              "type": "",
+              "text": "",
+              "emoji": false
+            },
+            "value": ""
+          },
+          {
+            "type": "",
             "fallback": "",
             "text": {
               "type": "",
