+++ json-logs/samples/api/conversations.history.json
@@ -29,11 +29,12 @@
         ],
         "replies": [
           {
-            "user": "B00000000",
-            "ts": "0000000000.000000"
+            "user": "",
+            "ts": ""
           }
         ],
-        "subscribed": false
+        "subscribed": false,
+        "last_read": "0000000000.000000"
       },
       "username": "",
       "icons": {
@@ -50,8 +51,8 @@
       ],
       "replies": [
         {
-          "user": "B00000000",
-          "ts": "0000000000.000000"
+          "user": "",
+          "ts": ""
         }
       ],
       "subscribed": false,
@@ -101,6 +102,94 @@
           "video_html": "",
           "video_html_width": 12345,
           "video_html_height": 12345,
+          "footer": "",
+          "footer_icon": "",
+          "ts": "",
+          "mrkdwn_in": [
+            ""
+          ],
+          "actions": [
+            {
+              "id": "",
+              "name": "",
+              "text": "",
+              "style": "",
+              "type": "",
+              "value": "",
+              "confirm": {
+                "title": "",
+                "text": "",
+                "ok_text": "",
+                "dismiss_text": ""
+              },
+              "data_source": "",
+              "min_query_length": 12345,
+              "url": ""
+            }
+          ],
+          "filename": "",
+          "size": 12345,
+          "mimetype": "",
+          "url": "",
+          "metadata": {
+            "thumb_64": false,
+            "thumb_80": false,
+            "thumb_160": false,
+            "original_w": 12345,
+            "original_h": 12345,
+            "thumb_360_w": 12345,
+            "thumb_360_h": 12345,
+            "format": "",
+            "extension": "",
+            "rotation": 12345,
+            "thumb_tiny": ""
+          }
+        },
+        {
+          "service_name": "",
+          "service_url": "https://www.example.com/",
+          "title": "",
+          "title_link": "https://www.example.com/",
+          "author_name": "",
+          "author_link": "https://www.example.com/",
+          "thumb_url": "https://www.example.com/",
+          "thumb_width": 12345,
+          "thumb_height": 12345,
+          "fallback": "",
+          "video_html": "",
+          "video_html_width": 12345,
+          "video_html_height": 12345,
+          "from_url": "https://www.example.com/",
+          "service_icon": "https://www.example.com/",
+          "id": 12345,
+          "original_url": "https://www.example.com/",
+          "msg_subtype": "",
+          "callback_id": "",
+          "color": "",
+          "pretext": "",
+          "author_icon": "",
+          "author_subname": "",
+          "channel_id": "",
+          "channel_name": "",
+          "bot_id": "",
+          "indent": false,
+          "is_msg_unfurl": false,
+          "is_reply_unfurl": false,
+          "is_thread_root_unfurl": false,
+          "is_app_unfurl": false,
+          "app_unfurl_url": "",
+          "text": "",
+          "fields": [
+            {
+              "title": "",
+              "value": "",
+              "short": false
+            }
+          ],
+          "image_url": "",
+          "image_width": 12345,
+          "image_height": 12345,
+          "image_bytes": 12345,
           "footer": "",
           "footer_icon": "",
           "ts": "",
@@ -443,6 +532,42 @@
           "external_id": "",
           "external_url": "https://www.example.com/",
           "has_rich_preview": false
+        },
+        {
+          "id": "F00000000",
+          "created": 12345,
+          "timestamp": 12345,
+          "name": "",
+          "title": "",
+          "mimetype": "",
+          "filetype": "",
+          "pretty_type": "",
+          "user": "U00000000",
+          "editable": false,
+          "size": 12345,
+          "mode": "",
+          "is_external": false,
+          "external_type": "",
+          "is_public": false,
+          "public_url_shared": false,
+          "display_as_bot": false,
+          "username": "",
+          "url_private": "https://www.example.com/",
+          "url_private_download": "https://www.example.com/",
+          "thumb_64": "https://www.example.com/",
+          "thumb_80": "https://www.example.com/",
+          "thumb_360": "https://www.example.com/",
+          "thumb_360_w": 12345,
+          "thumb_360_h": 12345,
+          "thumb_160": "https://www.example.com/",
+          "image_exif_rotation": 12345,
+          "original_w": 12345,
+          "original_h": 12345,
+          "thumb_tiny": "",
+          "permalink": "https://www.example.com/",
+          "permalink_public": "https://www.example.com/",
+          "is_starred": false,
+          "has_rich_preview": false
         }
       ],
       "upload": false,
@@ -465,7 +590,8 @@
         "team_id": "T00000000"
       },
       "topic": "",
-      "purpose": ""
+      "purpose": "",
+      "last_read": "0000000000.000000"
     }
   ],
   "has_more": false,
