+++ json-logs/samples/api/groups.replies.json
@@ -16,8 +16,8 @@
       ],
       "replies": [
         {
-          "user": "B00000000",
-          "ts": "0000000000.000000"
+          "user": "",
+          "ts": ""
         }
       ],
       "subscribed": false
