+++ json-logs/samples/api/mpim.replies.json
@@ -18,8 +18,8 @@
       ],
       "replies": [
         {
-          "user": "B00000000",
-          "ts": "0000000000.000000"
+          "user": "",
+          "ts": ""
         }
       ],
       "subscribed": false
