+++ json-logs/samples/api/stars.list.json
@@ -223,8 +223,8 @@
         ],
         "replies": [
           {
-            "user": "U00000000",
-            "ts": "0000000000.000000"
+            "user": "",
+            "ts": ""
           }
         ],
         "subscribed": false,
