+++ json-logs/samples/api/team.accessLogs.json
@@ -8,7 +8,10 @@
       "date_last": 12345,
       "count": 12345,
       "ip": "",
-      "user_agent": ""
+      "user_agent": "",
+      "isp": "",
+      "country": "",
+      "region": "12345"
     }
   ],
   "paging": {
