+++ json-logs/samples/rtm/PinRemovedEvent.json
@@ -545,12 +545,15 @@
       "preview_plain_text": "",
       "has_more": false,
       "sent_to_self": false,
-      "bot_id": "",
       "lines": 123,
       "lines_more": 123,
       "is_public": false,
       "public_url_shared": false,
       "display_as_bot": false,
+      "shares": {},
+      "channel_actions_ts": "",
+      "channel_actions_count": 123,
+      "bot_id": "",
       "initial_comment": {
         "id": "",
         "created": 123,
@@ -562,10 +565,7 @@
       },
       "num_stars": 123,
       "is_starred": false,
-      "comments_count": 123,
-      "channel_actions_ts": "",
-      "channel_actions_count": 123,
-      "shares": {}
+      "comments_count": 123
     },
     "comment": {
       "id": "",
